feat: Fix secondary button styling
Secondary small and large padding had been removed because it should be
the same as the standard button's large and small padding. However, the
removed specifcity caused bugs where the Button padding was getting
overridden by component styles, such as padding: 0 in buttonReset,
and caused buttons such as the CodeCTA button to lose their padding.